Split out normalizeTermsForApi
